Get named colours vector

Get named colours vector

Usage

get_colour(
  pattern = NULL,
  model = c("hex", "rgb", "cmyk", "pantone"),
  named = FALSE
)

get_colours(
  pattern = NULL,
  model = c("hex", "rgb", "cmyk", "pantone"),
  named = FALSE
)

Arguments

pattern

Optional. A character value or vector to use as a search term. Default is NULL in which case all the Oxford colours are returned.

model

A character vector of colour model. Can be "rgb", "cmyk", "hex", or "pantone". Default is "hex".

named

Logical. Should the output be a named character value or vector? Default is FALSE.

Value

A character value or vector of colour/s as per model specification. If named is TRUE, returns a named character value or vector.

Examples

get_colours()
#>  [1] "#348F41" "#B4A269" "#000000" "#6B4C24" "#E08F2A" "#C45B39" "#782C2D"
#>  [8] "#4B5430" "#002244" "#009FDA" "#000000" "#FFFFFF" "#F05023" "#FDB714"
#> [15] "#EB1C2D" "#F78D28" "#009CA7" "#00AB51" "#872B90" "#00A996" "#98252B"
#> [22] "#E16A2D" "#B88C1D" "#614776" "#006068" "#006450"
get_colours(model = "rgb")
#>  [1] "52, 143, 65"   "180, 162, 105" "0, 0, 0"       "107, 76, 36"  
#>  [5] "224, 143, 42"  "196, 91, 57"   "120, 44, 45"   "75, 84, 48"   
#>  [9] "0 35 69"       "0 173 228"     "0, 0, 0"       "225, 225, 225"
#> [13] "240, 80, 35"   "253, 183, 20"  "235, 28, 45"   "247, 141, 40" 
#> [17] "0, 156, 167"   "0, 171, 81"    "135, 43, 144"  "0, 169, 150"  
#> [21] "152, 37, 43"   "225, 106, 45"  "184, 140, 29"  "97, 71, 118"  
#> [25] "0, 96, 104"    "0, 100, 80"   
get_colours(pattern = "orange")
#> character(0)
get_colours(pattern = c("orange", "brown"), named = TRUE)
#> character(0)