Colours based on visual identity guidelines of various organisations

Colours based on visual identity guidelines of various organisations

Usage

paleta_colours

Format

A tibble with 5 fields

Variable	Description

name	Organisation colour name

code	Organisation colour code

rgb	Three integers for the red, green, blue components of the RGB colour model

cmyk	Four integers for the cyan, magenta, yellow, and black components of the CMYK colour model

hex	Hexadecimal codes for corresponding colour

pantone	Pantone colour name

Examples

paleta_colours
#> # A tibble: 60 × 7
#>    organisation name             code            rgb         cmyk  hex   pantone
#>  * <chr>        <chr>            <chr>           <chr>       <chr> <chr> <chr>  
#>  1 Africa CDC   ACDC Green       acdc_green      52, 143, 65 81, … #348… 7740C  
#>  2 Africa CDC   ACDC Gold        acdc_gold       180, 162, … 31, … #B4A… 4515C  
#>  3 Africa CDC   ACDC Black       acdc_black      0, 0, 0     0, 0… #000… Black C
#>  4 Africa CDC   ACDC Brown       acdc_brown      107, 76, 36 43, … #6B4… 1405C  
#>  5 Africa CDC   ACDC Yellow      acdc_yellow     224, 143, … 11, … #E08… 129C   
#>  6 Africa CDC   ACDC Orange      acdc_orange     196, 91, 57 18, … #C45… 158C   
#>  7 Africa CDC   ACDC Red         acdc_red        120, 44, 45 33, … #782… 1807C  
#>  8 Africa CDC   ACDC Dark Green  acdc_dark_green 75, 84, 48  69, … #4B5… 364C   
#>  9 World Bank   WB Sapphire Blue wb_blue         0 35 69     96, … #002… 5003C  
#> 10 World Bank   WB Cyan          wb_cyan         0 173 228   100,… #009… 5015C  
#> # ℹ 50 more rows